GroupConstants.   Weened core off GroupTypeEnum usage

+++ Mutagen.Bethesda.Tests/Passthrough Tests/OblivionPassthroughTest.cs
@@ -161,7 +161,7 @@ namespace Mutagen.Bethesda.Tests
                 new RecordType("DATA"),
             };
             ret.SetGroupAlignment(
-                GroupTypeEnum.CellTemporaryChildren,
+                (int)GroupTypeEnum.CellTemporaryChildren,
                 new RecordType("LAND"),
                 new RecordType("PGRD"));
             return ret;
